+++ header.php
@@ -25,7 +25,7 @@
 <div id="page" class="site">
 	<a class="skip-link screen-reader-text" href="#primary"><?php esc_html_e( 'Skip to content', '_s' ); ?></a>
 
-	<header id="masthead" class="site-header">
+	<header id="masthead" class="site-header"><div>
         <div class="site-branding">
 			<?php the_custom_logo();
 
@@ -47,8 +47,8 @@
 			'echo'           => false
 		] ) ) : ?>
             <nav id="site-navigation" class="main-navigation">
-                <button class="menu-toggle" aria-controls="navigation" aria-expanded="false"><?php esc_html_e( 'Menu', 'telly' ); ?></button>
+                <button class="menu-toggle" aria-controls="navigation" aria-expanded="false"><?php esc_html_e( 'Menu', '_s' ); ?></button>
 				<?php echo $navigation; ?>
             </nav><!-- #site-navigation -->
 		<?php endif; ?>
-	</header><!-- #masthead -->
+    </div></header><!-- #masthead -->